requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
pillow>=10.2.0
numpy>=1.26.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiofiles>=23.0.0
